requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
pypdf2
python-dotenv
numpy
pymupdf
sentence-transformers